requirements.txt
selenium
webdriver-manager
python-dotenv 
requests